add handling for frus-history/events page
note: because the links to events are hard-coded in the "frus-history" repo, the links don't adapt to the eXist app URL space (/exist/apps). so these links break when installed in a /exist/apps-based URL space. not currently worth the effort to re-work the links to work. we should also be aware of other places where this same phenomenon might be present.