fix(odd): Display frus-images in their natural size
Delete width and height for frus-images from odd.
This prevents images from being displayed in wrong ratios, when given
incorrect width and height values as inline styles in the TEI.
Fixes #115

TODO: Store the images size as metadata and use them to set width and
height dynamically, because these attributes are important for valid
markup.