fix (odd): Add rule for styling a nested list
This is a temporary fix open for discussion. I tried to style a list
that is nested within a list-item. In css, the selector I try to
describe is ul > ul.
With this fix, both lists get the same styling, so I guess the rule is
different here in odd.