Fix problem of missing volume titles in search page. Closes #251.
- It appears that templating directives that return attribute nodes has
the effect of “eating” any child elements, such as the child span
element here that was supposed to produce the volume title.
- The fix is to move attribute-emitting data-templates into its own
span element.
- Not sure if this behavior is a bug or by design, but anytime I have
to use templating to add one or more attributes to an element in the
HTML template, I use a child span element to do this, rather than
tacking the data template onto the target element itself.