First steps at addressing principal officer terms of service
- career ambassador is a permanent marker of distinguished service, so
we do not imply an end date
- some principal officer positions, like career ambassador, have only
dates of appointment and not (separate) dates of entry on duty; so when
determining a term of service, we would naturally start from the date
of appointment. but we do not have an explicit list of which principal
officer positions fall into which category. we’ll defer general changes
to the handling of terms of service summary until we’re able to confer
with Evan Duncan. In advance of that discussion we will prepare a
report profiling the different principal officer positions
- In the interim, the only position we’re changing terms of service
calculations for is career ambassador - and we’re starting from
appointment.  e.g., see Douglas MacArthur at
https://history.state.gov/departmenthistory/people/macarthur-douglas
and https://history.state.gov/departmenthistory/people/by-name/m