New POCOM database
Joe: now that the next text is available, is there a way to ask
Stephanie or another one of the editors to take a look at it and give
it an editorial pass?

Per your request, I have loaded this text into GitHub but not uploaded
it to the website.