(feat)vols filter shows only vols with hits
On search results page, the Volumes filter should only show the volumes that have hits for the requested search. Thus, the search:load-results template function call has to move up to be surrounding the filters and results. And the search:select-volumes function should be able to draw from the list of volume-ids from the $hits, so we add a new  “results-doc-ids” entry to the “query-info” map.