Fix broken footnote links, closes #258
Catch requests for footnote IDs, redirect to the parent document (via a 302) and pass the appropriate fragment identifier for the requested target - so that the browser goes straight to the footnote.

My earlier attempt (#263) attempted to pre-process links while rendering documents, to generate the correct link—ensuring “correctness” by looking first to see if the target actually existed or not—but this proved to be a taxing operation.

This new approach does the opposite: we now wait for a footnote link to be requested, and then (re-)route the request as needed. This should impose negligible overhead.

Closes #258 and #263.