fix(scss): Add default width for images in social media posts
On small screens the image will expand to the maximum width of the
wrapping element, which is in this case the full screen width,
resulting in small images to scale up to the full width of the screen.
Before, the image expanded to its own original width, which resulted in
different image sizes, sometimes even smaller than the wrapping
container.