refactor(templates): `pages:load` runs first
Ensuring `pages:load` or similar are run before wrapping the page in the site template means that any information added to `$model` is available when e.g. generating metadata in the HTML head.
NB: although this commit is in itself a refactor, it is required to support planned features such as social media cards/Open Graph metadata generation.