fix(breadcrumbs): Use only  selected URI paramters
Using any URI parameters supplies unwanted parameters (such as the section-id) to the pages:generate-breadcrumb-item() function; this causes problems e.g. when returning the breadcrumb title for the document level (which instead returns the section title).
One solution for this would be to provide a list of URI parameters for each URL, or to specifically exclude certain parameters.  Instead I have chosen the simpler way of only allowing certain named parameters through.