fix: account for fn:root bugfix in eXist 6.1.0
Before eXist 6.1.0, eXist’s implementation of the fn:root() function contained a bug that caused it to always return a document node, even when called on an in-memory node whose root node was an element node. This bug was fixed in https://github.com/eXist-db/exist/pull/4402.

The only code in hsg-shell that assumed the buggy behavior was the code that handled the display of concurrent appointments in POCOM data.

After upgrading HSG from 5.1.1 to 6.2.0, this code triggered an error.